Import all pkg which tests depend on
All imports which gazelle identified as test dependencies should be
imported in the test main package.

This way "hidden" global registrations are properly added to the binary.

One use-case is Ginkgo. It has its own wrappers for tests and rules_go
would miss them in some cases, which would lead to tests which are not
added to the binary.

Signed-off-by: Roman Mohr <rmohr@redhat.com>