Remove reference to platforms repository
It is better to use version that comes with bazel, because no version
bumping is needed.

Refering to a fixed version resulted in two broken downstream project:
FlatBuffers and Cloud Robotics core, which depended on rules_go version
0.20.3, which depends on an old version of platforms, whcih is missing
definitions of macos os and arm64 cpus, whcih is needed by the latest
bazel release.