link: make package conflicts into errors instead of warnings
We've promised to do this for more than two years, and the time has
finally come.

If more than one copy of a package with the same package path is
provided to the linker, the GoLink action now reports an
error. Distinct packages (for example, from multiple vendor
directories) may be disambiguated with the importmap attribute.

Fixes #1374