go_proto_library: add implicit dependencies on APIv2 packages
Since protobuf 3.14, the Well Known Types declare APIv2 package paths,
which means generated code that imports them will import APIv2
packages.

This PR is intended to be cherry-picked to the 0.23 and 0.24 release
branches. It may slow down builds, since both APIv1 and APIv2 packages
will be compiled; we can't tell which is needed during analysis.

For #2721