Add initial bzlmod setup
This commit uses MODULE.bazel to load all regular dependencies that are
already available in the Bazel Central Registry. The contents of the
WORKSPACE file will be used as a fallback, so dependencies can be
"modularized" incrementally.

Also adds a Bazel CI job to verify the module setup.

Note: Until all repositories in repositories.bzl are loaded from
MODULE.bazel, rules_go can only be used as the main module. The next
step thus consists of refactoring repositories.bzl such that it can be
used both as a WORKSPACE macro and a module extension.