Use target package when constructing pkg.json filename
Use the package name in the pkg.json for each dependency in a
particular target when processing a request from gopls. This ensures
that separate packages that have the same package name but different
import paths do not collide and that each import map to a unique
pkg.json file during the analysis step.

Closes #2948