Allow patching the toolchain
This change adds a patch and patch_strip attribute to the
go_download_sdk rule. This is only available when calling this rule
directly, and not when using the go_register_toolchains wrapper.

This allows for the applicable of a single patch to the toolchain, using
the built-in repository_ctx patch method.