sdk: symlink everything under lib/ when using local SDK
We rely on the packaged zoneinfo zip in CRDB, found under
lib/time/zoneinfo.zip. When using the remote SDK macro, we untar the
entire archive into the sandbox. When pointing to a local SDK however,
we previously ignored everything under lib. It feels reasonable to also
symlink lib/ into the sandbox given they have valid symbols to build
against.

We needed this patch in order to build CRDB with local go installations.

+cc https://github.com/cockroachdb/cockroach/issues/56178