Remove unnecessary rpath special handling on macOS
Based on
https://github.com/bazelbuild/bazel/issues/10254#issuecomment-1125218345
and https://github.com/bazelbuild/bazel/pull/12304 being fixed, this
special handling of rpaths on macOS appears to be unnecessary.

This cleanup ensures that Bazel sets the correct metadata for the exec
location of Go libraries linked in c-shared mode, which in turn allows
to not include them in the runfiles of all dependents - cc_* targets
depending on them will now generate the correct rpath entries to find
the libraries at runtime at the usual position.