go_path: Support go:embed
The `go_path` rule does not currently copy embedded sources.
This means that if we use `go:embed` in any file and use `go_path` to
build a GOPATH from that, the resulting code will not compile.
Example of an error from one of our internal systems:

    scaffold.go:9:13: pattern template/*: no matching files found

To fix this, the go_path rule needs to copy the embedsrcs over from the
archive and add them to the manifest for the go_path builder to copy.