Handle embedsrcs inside subdirectories
The previous implementation of this did not handle a `go:embed` for a
subdirectory that was not a Bazel package. For example,

    mypkg/
    |- BUILD.bazel
    |- foo.go   // contains "go:embed template/foo"
    |- template/
        |- foo

With this, the BUILD.bazel for mypkg will hold
`embedsrcs = ["template/foo"]`.

However, if we just append `f.basename` to the output directory path,
this will generate `mypkg/foo`, not `mypkg/template/foo`.

To fix this, this commit is tracking the directory in which the last
source file was found,
and use a path to the embedded file relative to that directory.