Link in native libraries of transitive dependencies in archive mode
When linking a go_binary in one of the native archive modes, all
native libraries of transitive dependencies should be included in the
resulting CcInfo, not just the libraries obtained from the cdeps.

This commit adds a `cc_info` field to the `GoSource` provider that
includes the result of merging the `CcInfo`s of all `deps` and `cdeps`
for all Go libraries, not just for those using cgo themselves.