Add support for `go_binary` to gopackagesdriver.
As discussed in #3270, The current implementation of `bazel_json_builder`
does not handle `go_binary` build rules. This means that when
`gopackagesdriver` attempts to be used on a file that is apart of a
`go_binary`, no build target is produced as the output of `bazel query`.
This prevents downstream utilities (`gopls`) from being able to provide
functionality such as autocompletion support.

This PR simply adds `go_binary` as a supported rule type, fixing #3270.