always produce .a files at the beginning of a build
This change updates rules_go in anticipation for Go 1.20, which will
ship without .a files for the standard library. Now, instead of
conditionally building the standard library if the configuration
differs from the default, the standard library will always be built.
In addition, the builder will be built using 'go build' instead of
individual invocations of go tool compile/link to avoid needing to
bootstrap the standard library before producing the builder.