Only infer platforms from crosstool/cpu if at default
If the current target platform differs from the default value (the host
platform), then we don't want to override it with a value inferred from
the legacy --cpu and --crosstool_top flags. Otherwise it would become
impossible to "platformize" a Go build without having a matching
platform mappings file.

This is not expected to break any existing use cases and if it does, it
can and should be worked around with a platform mappings file that maps
the legacy flag values to `//go/platform` targets.