Keep build tags that affect the stdlib build
Since 8d309d5, all `gotags` were cleared from the configuration when
building the standard library. This is not correct as some tags, such
as `timetzdata`, do affect the standard library build.

This commit adds a tool that scrapes the list of build tags relevant
for the standard library from Go SDK sources and keeps those tags that
are relevant for Go 1.18-1.20.