Disable cgo by default with unsupported compilers
MSVC-style compilers are explicitly not supported by cgo. If the
current C/C++ toolchain is detected to be of that type, disable cgo by
default unless explicitly requested.

Without this commit, building the standard library on Windows fails even
with #3491 without forcing use of mingw-gcc, which we don't want to ask
of our Go-only users.