Remove tests for `go_path`'s `link` mode
`link` mode produces non-hermetically dangling symlinks by design, which
means that it is broken with remote caching or execution and cannot be
guaranteed to work in other situations.

In the feature, we may be able to support it again based on
`ctx.declare_symlink` and improved unresolved symlink support with
remote caching and execution.