cgo: Control static linking through rules_go
Let the link action of rules_go determine how to link the binary. This
aligns the behavior of the stdlib with other cgo code.

A binary can be linked statically by either using `static = "on"` in
the go_binary rule or using `--@io_bazel_rules_go//go/config:static`.

Fixes #3601