Allow generating export files for the stdlib.
By default, rules_go only generates Export files for the code it
compiles using bazel, but not the stdlib. This is problematic for tools
like golangci-lint that has analyzers that utilize the export files for
their checks, including the stdlib.

It's possible to generate export files for the stdlib by passing the
`-export` flag to the `go list` command, however, it comes with a time
penalty (about 20s with the flag vs 2s without the flag).

Therefore, add it as an option for go_register_toolchains.

Now, you can do something like:
```
go_register_toolchains(
    exports_for_stdlib = True,
    version = "1.20.1",
)
```

And it will generate the export files in the stdlib.pkg.json. Note that
this time penalty only seems to be when using gopackagedriver not for
regular builds.

Fixes #3651