go/tools/gopackagesdriver: pass Compiler and Arch in DriverResponse
go/packages is being changed to expect Compiler and Arch in its
DriverResponse (see golang.org/cl/516917) because we can't expect the
type of the types.Sizes returned by types.Sizes to be types.StdSizes.
(The default it uses when Compiler and Arch aren't set is the set of
types for gc,amd64, so there's no change in behavior if those fields
aren't set, but we should set them. I'd also like to see if we can
correctly provide the arch assuming it's not amd64. I left a question
in the code asking about that.