Fix hermetic toolchain linking before Go 1.21
Go toolchain checks which linker flags are available by running the
linker on a temporary file. Before Go 1.21, this command was run in a
temporary directory, but this has now been fixed upstream.
https://github.com/golang/go/issues/59952

This change will not be needed for Go versions 1.21 and above, but will
be beneficial for users on lower versions and using a hermetic
toolchain.

See https://github.com/grailbio/bazel-toolchain/issues/183 for some more
context.