make toolchains repo more reproducible
The versions.json file results from downloading a list of versions from go.dev.
(https://go.dev/dl/?mode=json)
This API is meant to be updated for every Go release and will result in different outputs when rerun in the future.

By deleting the file, we make the behavior more reproducible: As long as a specific version is requested, this should result in reproducible behavior.
(This assumes that go.dev does not replace released SDK artifacts, which should be a fair assumption).